requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
anytree==2.12.1
asttokens==2.4.1
attrs==24.2.0
autograd==1.7.0
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.17.1
azure-search-documents==11.5.1
azure-storage-blob==12.22.0
beartype==0.18.5
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cramjam==2.8.3
cryptography==43.0.1
cycler==0.12.1
dask==2024.9.0
dask-expr==1.1.14
datashaper==0.0.49
decorator==5.1.1
deprecation==2.1.0
devtools==0.12.2
diskcache==5.6.3
distro==1.9.0
environs==11.0.0
executing==2.1.0
fastapi==0.112.4
fastparquet==2024.5.0
ffmpy==0.4.0
filelock==3.16.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
gensim==4.3.3
gradio==4.43.0
gradio_client==1.3.0
graspologic==3.4.1
graspologic-native==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
hyppo==0.4.0
idna==3.9
importlib_resources==6.4.5
ipython==8.27.0
ipywidgets==8.1.5
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lancedb==0.10.2
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
ollama==0.3.3
openai==1.45.0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pillow==10.4.0
plotly==5.24.1
portalocker==2.10.1
POT==0.9.4
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
py==1.11.0
pyaml-env==1.2.1
pyarrow==15.0.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.14.1
pynndescent==0.5.13
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
pywin32==306
PyYAML==6.0.2
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
retry==0.9.2
rich==13.8.1
rpds-py==0.20.0
ruff==0.6.5
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
setuptools==74.1.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.5
statsmodels==0.14.2
swifter==1.4.0
tabulate==0.9.0
tenacity==8.5.0
textual==0.72.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
umap-learn==0.5.6
urllib3==2.2.3
uvicorn==0.30.6
wcwidth==0.2.13
websockets==12.0
widgetsnbextension==4.0.13
wrapt==1.16.0
yarl==1.11.1
yfiles_jupyter_graphs==1.8.1
